refactor: aleterado arquivo style para styles
