+++ src/components/molecules/NotificationsMenu/styles.ts
